moved functions to handle the loading of Saved_models, and extraction of graph_defs from 'aot.py' to 'tools.py'
